+++ src/main/java/com/dabomstew/pkrandom/romhandlers/Gen5RomHandler.java
@@ -1169,6 +1169,7 @@ public class Gen5RomHandler extends AbstractDSRomHandler {
             getTextHandler().bw1CherenBurghTextModifications(taggedTypes);
             getTextHandler().bw1NimbasaCityTextModifications(taggedTypes, this.random);
             getTextHandler().bw1JuniperTextModifications(taggedTypes);
+            getTextHandler().bw1DriftveilCityTextModifications(taggedTypes, this.random);
         }
     }