+++ src/test/java/com/dabomstew/pkrandom/romhandlers/Gen5Test.java
@@ -240,6 +240,30 @@ public class Gen5Test {
     }
 
     /**
+     * When altering trainers, make sure the ROM's text is altered when it is a
+     * Black/White ROM
+     * @throws IOException
+     */
+    @Test
+    public void TestBW1DriftveilCityText() throws IOException {
+        // Test with and without null
+        Gen5RomHandler romhandler = spy(new Gen5RomHandler(new Random()));
+        resetDataModel(romhandler);
+        doReturn(Gen5RomHandler.getRomFromSupportedRom("Black (U)")).when(romhandler).getRomEntry();
+        doReturn(0xFFFF).when(romhandler).readWord(any(), anyInt());
+        doReturn(pokemonList.subList(0, 3)).when(romhandler).getStarters();
+        
+        // Verify with random team
+        romhandler.randomizeTrainerPokes(false, false, false, 0);
+        verify(mockTextHandler, times(1)).bw1DriftveilCityTextModifications(isNull(), any());
+
+        // Verify with type theming
+        romhandler.typeThemeTrainerPokes(false, false, false, false, false, 0);
+        verify(mockTextHandler, times(1)).bw1DriftveilCityTextModifications(isNotNull(), any());
+    }  
+    
+
+    /**
      * Function for granular modification of data model
      */
     private void setUp() {